test: fix some storybook stories, rearrange Components stories hierarchy
