+++ typescript/packages/subsurface-viewer/src/storybook/components/colorLegends/ContinuousLegend.stories.tsx
@@ -1,32 +1,34 @@
 import React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+
+import type { colorTablesArray } from "@emerson-eps/color-tables";
 import { ContinuousLegend, colorTables } from "@emerson-eps/color-tables";
-export default {
+
+const stories: Meta = {
     component: ContinuousLegend,
-    title: "SubsurfaceViewer/Components/ColorLegends/ContinuousLegend",
+    title: "SubsurfaceViewer/Components/ColorLegends",
 };
+export default stories;
 
 const min = 0;
 const max = 0.35;
 const dataObjectName = "Wells / PORO";
-const position = [16, 10];
 const name = "PORO";
 const horizontal = false;
 const colorName = "Rainbow";
 const reverseRange = false;
 
-const Template = (args) => {
-    return <ContinuousLegend {...args} />;
-};
-
-export const ContinuousTemplate = Template.bind({});
-ContinuousTemplate.args = {
-    min,
-    max,
-    dataObjectName,
-    position,
-    name,
-    colorName,
-    colorTables,
-    horizontal,
-    reverseRange,
+export const ContinuousLegendStory: StoryObj<typeof ContinuousLegend> = {
+    name: "ContinuousLegend",
+    args: {
+        min,
+        max,
+        dataObjectName,
+        id: name,
+        colorName,
+        colorTables: colorTables as colorTablesArray,
+        horizontal,
+        reverseRange,
+    },
+    render: (args) => <ContinuousLegend {...args} />,
 };
