+++ typescript/packages/subsurface-viewer/src/storybook/components/colorLegends/IndividualScaleForMap.stories.tsx
@@ -1,11 +1,18 @@
 import React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { View } from "@deck.gl/core/typed";
+
+import { ColorLegend } from "@emerson-eps/color-tables";
+
 import exampleData from "../../../../../../../example-data/deckgl-map.json";
 import SubsurfaceViewer from "../../../SubsurfaceViewer";
 
-export default {
+const stories: Meta = {
     component: SubsurfaceViewer,
-    title: "SubsurfaceViewer/Components/ColorLegends/IndividualScaleForMap",
+    title: "SubsurfaceViewer/Components/ColorLegends",
 };
+export default stories;
 
 // Template for when edited data needs to be captured.
 const EditDataTemplate = (args) => {
@@ -20,80 +27,137 @@ const EditDataTemplate = (args) => {
             setProps={(updatedProps) => {
                 setEditedData(updatedProps.editedData);
             }}
-        />
+        >
+            <>
+                {
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                    /* @ts-expect-error */
+                    <View id="view_1">
+                        <ColorLegend
+                            min={3000}
+                            max={3100}
+                            horizontal={true}
+                            dataObjectName={"Rainbow"}
+                            colorName={"Rainbow"}
+                        />
+                    </View>
+                }
+                {
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                    /* @ts-expect-error */
+                    <View id="view_2">
+                        <ColorLegend
+                            min={3000}
+                            max={3100}
+                            horizontal={true}
+                            dataObjectName={"Porosity"}
+                            colorName={"Porosity"}
+                        />
+                    </View>
+                }
+                {
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                    /* @ts-expect-error */
+                    <View id="view_3">
+                        <ColorLegend
+                            min={3000}
+                            max={3100}
+                            horizontal={true}
+                            dataObjectName={"Permeability"}
+                            colorName={"Permeability"}
+                        />
+                    </View>
+                }
+                {
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                    /* @ts-expect-error */
+                    <View id="view_4">
+                        <ColorLegend
+                            min={3000}
+                            max={3100}
+                            horizontal={true}
+                            dataObjectName={"Seismic"}
+                            colorName={"Seismic"}
+                        />
+                    </View>
+                }
+            </>
+        </SubsurfaceViewer>
     );
 };
 
 // 4 maps with individual color scales for each map
-export const IndividualScaleForMap = EditDataTemplate.bind({});
-IndividualScaleForMap.args = {
-    ...exampleData[0],
-    legend: {
-        visible: true,
-    },
-    zoom: -5,
-    layers: [
-        exampleData[0].layers[0],
-        {
-            ...exampleData[0].layers[0],
-            colorMapRange: [3000, 3100],
-            valueRange: [3000, 3100],
-            id: "colormap-2-layer",
-            colorMapName: "Porosity",
-        },
-        {
-            ...exampleData[0].layers[0],
-            colorMapRange: [3000, 3100],
-            valueRange: [3000, 3100],
-            id: "colormap-3-layer",
-            colorMapName: "Permeability",
-        },
-        {
-            ...exampleData[0].layers[0],
-            colorMapRange: [3000, 3100],
-            valueRange: [3000, 3100],
-            id: "colormap-4-layer",
-            colorMapName: "Seismic",
-        },
-    ],
-    views: {
-        layout: [2, 2],
-        showLabel: true,
-        viewports: [
+export const IndividualScaleForMap: StoryObj<typeof SubsurfaceViewer> = {
+    name: "Individual ContinuousLegend",
+    args: {
+        ...exampleData[0],
+        layers: [
             {
-                id: "view_1",
-                name: "Colormap 1 layer",
-                show3D: false,
-                layerIds: ["colormap-layer"],
+                ...exampleData[0].layers[0],
+                id: "colormap-layer",
             },
             {
-                id: "view_2",
-                name: "Colormap 2 layer",
-                show3D: false,
-                layerIds: ["colormap-2-layer"],
+                ...exampleData[0].layers[0],
+                colorMapRange: [3000, 3100],
+                valueRange: [3000, 3100],
+                id: "colormap-2-layer",
+                colorMapName: "Porosity",
             },
             {
-                id: "view_3",
-                name: "Colormap 3 layer",
-                show3D: false,
-                layerIds: ["colormap-3-layer"],
+                ...exampleData[0].layers[0],
+                colorMapRange: [3000, 3100],
+                valueRange: [3000, 3100],
+                id: "colormap-3-layer",
+                colorMapName: "Permeability",
             },
             {
-                id: "view_4",
-                name: "Colormap 4 layer",
-                show3D: false,
-                layerIds: ["colormap-4-layer"],
+                ...exampleData[0].layers[0],
+                colorMapRange: [3000, 3100],
+                valueRange: [3000, 3100],
+                id: "colormap-4-layer",
+                colorMapName: "Seismic",
             },
         ],
-    },
-};
-
-IndividualScaleForMap.parameters = {
-    docs: {
-        description: {
-            story: "Four maps with individual color scales for each map",
+        views: {
+            layout: [2, 2],
+            showLabel: true,
+            viewports: [
+                {
+                    id: "view_1",
+                    name: "Colormap 1 layer",
+                    show3D: false,
+                    layerIds: ["colormap-layer"],
+                },
+                {
+                    id: "view_2",
+                    name: "Colormap 2 layer",
+                    show3D: false,
+                    layerIds: ["colormap-2-layer"],
+                },
+                {
+                    id: "view_3",
+                    name: "Colormap 3 layer",
+                    show3D: false,
+                    layerIds: ["colormap-3-layer"],
+                },
+                {
+                    id: "view_4",
+                    name: "Colormap 4 layer",
+                    show3D: false,
+                    layerIds: ["colormap-4-layer"],
+                },
+            ],
         },
-        inlineStories: false,
-        iframeHeight: 500,
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } as any,
+    parameters: {
+        docs: {
+            description: {
+                story: "Four maps with individual color scales for each map",
+            },
+            inlineStories: false,
+            iframeHeight: 500,
+        },
     },
+    render: (args) => <EditDataTemplate {...args} />,
 };
